+++ changelog.md
@@ -1,5 +1,10 @@
+# 1.0.2
+- Support for harder sprite/frame names like full path or with relative subdir (mod.id/sprite.png)
+- Renamed `names` key to `frames` in animator setup
+- Added `spritesheets` for animator setup
+
 # 1.0.1
-- More animate data files paths list updates.
+More animate data files paths list updates
 
 # 1.0.0
-- First release.
+First release